Refactor: Improve PHP application structure and code quality
This commit introduces a major refactoring of the PHP application to improve its overall structure, maintainability, and security.

Key changes:
- **Directory Restructuring:** Project files have been organized into a more logical structure:
  - `public/`: New web root for all user-facing files and assets.
  - `src/`: Contains all backend PHP logic, including handlers and libraries, now located outside the web root for better security.
  - `views/`: Contains all HTML presentation files, separated into layouts and partials.
  - `lang/`: Contains language files for localization.

- **Separation of Concerns:** The monolithic PHP files (`login.php`, `admin.php`, `rsvp.php`) have been broken down. Each page now has:
  - A logic handler in `src/handlers/`.
  - A view template in `views/`.
  - A simple entry-point file in `public/` to bootstrap the application and render the view.

- **Centralized Bootstrap:** A new `src/bootstrap.php` file now handles common tasks like session initialization, database connection, and language loading.

- **Externalized Language Strings:** All hardcoded Arabic and English strings have been moved from the application files into `lang/ar.php` and `lang/en.php`.

- **Function Optimization:** Reusable functions like `safe_html`, `parseEventDateTime`, `formatDateArabic`, and a new `handle_upload` function have been created and moved to `src/lib/functions.php` to reduce code duplication.
