+++ lang/ar.php
@@ -1,0 +1,123 @@
+<?php
+// lang/ar.php
+
+return [
+    'login_title' => 'تسجيل الدخول',
+    'welcome_back' => 'مرحباً بعودتك',
+    'login_subtitle' => 'يرجى تسجيل الدخول للوصول إلى لوحة التحكم',
+    'username_label' => 'اسم المستخدم',
+    'username_placeholder' => 'أدخل اسم المستخدم',
+    'password_label' => 'كلمة المرور',
+    'password_placeholder' => 'أدخل كلمة المرور',
+    'remember_me' => 'تذكرني',
+    'login_button' => 'تسجيل الدخول',
+    'logging_in' => 'جاري تسجيل الدخول...',
+    'show_password' => 'إظهار كلمة المرور',
+    'hide_password' => 'إخفاء كلمة المرور',
+    'error_invalid_credentials' => 'اسم المستخدم أو كلمة المرور غير صحيحة.',
+    'error_account_locked' => 'تم قفل الحساب مؤقتاً. يرجى المحاولة بعد {minutes} دقائق.',
+    'error_too_many_attempts' => 'محاولات كثيرة جداً. يرجى الانتظار {seconds} ثانية.',
+    'error_fill_fields' => 'الرجاء إدخال اسم المستخدم وكلمة المرور.',
+    'error_csrf' => 'خطأ في التحقق من صحة الطلب. يرجى المحاولة مرة أخرى.',
+    'error_general' => 'عذراً، حدث خطأ ما. يرجى المحاولة مرة أخرى.',
+    'error_no_event_access' => 'هذا المستخدم غير مصرح له بالدخول.',
+    'attempts_remaining' => 'محاولات متبقية: {count}',
+    'login_success' => 'تم تسجيل الدخول بنجاح! جاري التحويل...',
+
+    // Admin Page
+    'administration' => 'الإدارة',
+    'logout' => 'تسجيل الخروج',
+    'event_settings' => 'إعدادات الحفل',
+    'user_management' => 'إدارة المستخدمين',
+    'event_details' => 'تفاصيل الحفل',
+    'event_name' => 'اسم الحفل (عربي)',
+    'event_name_en' => 'اسم الحفل (إنجليزي)',
+    'event_slug' => 'رابط الحفل المخصص (Slug)',
+    'google_maps_link' => 'رابط خرائط جوجل',
+    'event_date_time' => 'تاريخ ووقت الحفل',
+    'venue_ar' => 'مكان الحفل (عربي)',
+    'venue_en' => 'مكان الحفل (إنجليزي)',
+    'event_description_ar' => 'وصف الحفل (عربي)',
+    'event_description_en' => 'وصف الحفل (إنجليزي)',
+    'image_settings' => 'إعدادات الصور',
+    'display_image' => 'صورة العرض',
+    'whatsapp_image' => 'صورة الواتساب',
+    'current_display_image' => 'الصورة الحالية للعرض',
+    'current_whatsapp_image' => 'الصورة الحالية للواتساب',
+    'upload_display_image' => 'رفع صورة العرض',
+    'upload_whatsapp_image' => 'رفع صورة الواتساب',
+    'remove_current_image' => 'حذف الصورة الحالية',
+    'image_preview' => 'معاينة الصورة',
+    'cancel_selection' => 'إلغاء الاختيار',
+    'qr_settings' => 'إعدادات QR',
+    'webhook_settings' => 'إعدادات Webhook',
+    'registration_settings' => 'إعدادات صفحة التسجيل',
+    'registration_show_phone' => 'إظهار حقل الهاتف',
+    'registration_require_phone' => 'جعل الهاتف مطلوباً',
+    'registration_show_guest_count' => 'إظهار عدد الضيوف',
+    'registration_show_countdown' => 'إظهار العد التنازلي',
+    'registration_show_location' => 'إظهار معلومات الموقع',
+    'registration_mode' => 'نمط التسجيل',
+    'registration_mode_simple' => 'مبسط',
+    'registration_mode_full' => 'كامل',
+    'rsvp_settings' => 'إعدادات صفحة تأكيد الحضور (RSVP)',
+    'rsvp_show_guest_count' => 'إظهار حقل عدد الضيوف',
+    'rsvp_show_qr_code' => 'إظهار حقل رمز QR',
+    'save_all_settings' => 'حفظ جميع الإعدادات',
+    'image_saved_success' => 'تم حفظ الصورة بنجاح',
+    'image_removed_success' => 'تم حذف الصورة بنجاح',
+    'users' => 'المستخدمين',
+    'add_user' => 'إضافة مستخدم جديد',
+    'username' => 'اسم المستخدم',
+    'password' => 'كلمة المرور',
+    'role' => 'الدور',
+    'custom_event' => 'الحفل المخصص',
+    'add' => 'إضافة',
+    'current_users' => 'المستخدمون الحاليون',
+    'actions' => 'الإجراءات',
+    'edit' => 'تعديل',
+    'delete' => 'حذف',
+    'confirm_delete_user' => 'هل أنت متأكد من حذف هذا المستخدم؟',
+    'user_added_success' => 'تم إضافة المستخدم بنجاح.',
+    'user_add_error' => 'حدث خطأ في إضافة المستخدم.',
+    'user_updated_success' => 'تم تحديث المستخدم بنجاح.',
+    'user_update_error' => 'حدث خطأ في التحديث.',
+    'user_deleted_success' => 'تم حذف المستخدم بنجاح.',
+    'user_delete_error' => 'حدث خطأ في الحذف.',
+    'user_exists' => 'اسم المستخدم موجود بالفعل.',
+    'all_fields_required' => 'الرجاء إدخال كل الحقول.',
+    'leave_empty' => 'اتركه فارغاً لعدم التغيير',
+    'save_changes' => 'حفظ التعديلات',
+    'cancel' => 'إلغاء',
+
+    // RSVP Page
+    'welcome_guest' => 'مرحباً بكم',
+    'dear_guest' => 'ضيفنا الكريم',
+    'guest_count' => 'عدد الضيوف',
+    'table_number' => 'رقم الطاولة',
+    'confirm_attendance' => 'تأكيد الحضور',
+    'decline_attendance' => 'الاعتذار عن الحضور',
+    'add_to_calendar' => 'إضافة للتقويم',
+    'share_invitation' => 'مشاركة الدعوة',
+    'get_directions' => 'الحصول على الاتجاهات',
+    'download_qr' => 'تحميل QR',
+    'entry_card' => 'بطاقة الدخول',
+    'qr_code' => 'رمز الاستجابة السريعة',
+    'show_at_entrance' => 'أظهر هذا الرمز عند الدخول',
+    'already_confirmed' => 'تم تأكيد حضورك بنجاح!',
+    'already_declined' => 'تم تسجيل اعتذارك',
+    'success_confirmed' => 'تم تأكيد حضورك بنجاح!',
+    'success_declined' => 'شكراً لك، تم تسجيل اعتذارك عن الحضور.',
+    'error_occurred' => 'حدث خطأ، يرجى المحاولة مرة أخرى',
+    'invalid_link' => 'رابط الدعوة غير صالح',
+    'csrf_error' => 'خطأ في الحماية، يرجى إعادة تحميل الصفحة',
+    'rate_limit_error' => 'تم إرسال طلبات كثيرة، يرجى الانتظار',
+    'connection_error' => 'خطأ في الاتصال',
+    'countdown_title' => 'العد التنازلي للحفل',
+    'days' => 'يوم',
+    'hours' => 'ساعة',
+    'minutes' => 'دقيقة',
+    'seconds' => 'ثانية',
+    'event_time_reached' => '🎉 حان وقت الحفل! 🎉',
+    'enjoy_time' => 'نتمنى لكم وقتاً ممتعاً'
+];
