+++ lang/en.php
@@ -1,0 +1,123 @@
+<?php
+// lang/en.php
+
+return [
+    'login_title' => 'Login',
+    'welcome_back' => 'Welcome Back',
+    'login_subtitle' => 'Please sign in to access your dashboard',
+    'username_label' => 'Username',
+    'username_placeholder' => 'Enter your username',
+    'password_label' => 'Password',
+    'password_placeholder' => 'Enter your password',
+    'remember_me' => 'Remember me',
+    'login_button' => 'Sign In',
+    'logging_in' => 'Signing in...',
+    'show_password' => 'Show password',
+    'hide_password' => 'Hide password',
+    'error_invalid_credentials' => 'Invalid username or password.',
+    'error_account_locked' => 'Account temporarily locked. Please try again after {minutes} minutes.',
+    'error_too_many_attempts' => 'Too many attempts. Please wait {seconds} seconds.',
+    'error_fill_fields' => 'Please enter both username and password.',
+    'error_csrf' => 'Security token mismatch. Please try again.',
+    'error_general' => 'Sorry, something went wrong. Please try again.',
+    'error_no_event_access' => 'This user is not authorized to access.',
+    'attempts_remaining' => 'Attempts remaining: {count}',
+    'login_success' => 'Login successful! Redirecting...',
+
+    // Admin Page
+    'administration' => 'Administration',
+    'logout' => 'Logout',
+    'event_settings' => 'Event Settings',
+    'user_management' => 'User Management',
+    'event_details' => 'Event Details',
+    'event_name' => 'Event Name (Arabic)',
+    'event_name_en' => 'Event Name (English)',
+    'event_slug' => 'Custom Event Link (Slug)',
+    'google_maps_link' => 'Google Maps Link',
+    'event_date_time' => 'Event Date & Time',
+    'venue_ar' => 'Venue (Arabic)',
+    'venue_en' => 'Venue (English)',
+    'event_description_ar' => 'Event Description (Arabic)',
+    'event_description_en' => 'Event Description (English)',
+    'image_settings' => 'Image Settings',
+    'display_image' => 'Display Image',
+    'whatsapp_image' => 'WhatsApp Image',
+    'current_display_image' => 'Current Display Image',
+    'current_whatsapp_image' => 'Current WhatsApp Image',
+    'upload_display_image' => 'Upload Display Image',
+    'upload_whatsapp_image' => 'Upload WhatsApp Image',
+    'remove_current_image' => 'Remove Current Image',
+    'image_preview' => 'Image Preview',
+    'cancel_selection' => 'Cancel Selection',
+    'qr_settings' => 'QR Settings',
+    'webhook_settings' => 'Webhook Settings',
+    'registration_settings' => 'Registration Page Settings',
+    'registration_show_phone' => 'Show Phone Field',
+    'registration_require_phone' => 'Make Phone Required',
+    'registration_show_guest_count' => 'Show Guest Count Field',
+    'registration_show_countdown' => 'Show Countdown Timer',
+    'registration_show_location' => 'Show Location Info',
+    'registration_mode' => 'Registration Mode',
+    'registration_mode_simple' => 'Simple',
+    'registration_mode_full' => 'Full',
+    'rsvp_settings' => 'RSVP Page Settings',
+    'rsvp_show_guest_count' => 'Display Guest Count',
+    'rsvp_show_qr_code' => 'Display QR Code Field',
+    'save_all_settings' => 'Save All Settings',
+    'image_saved_success' => 'Image saved successfully',
+    'image_removed_success' => 'Image removed successfully',
+    'users' => 'Users',
+    'add_user' => 'Add New User',
+    'username' => 'Username',
+    'password' => 'Password',
+    'role' => 'Role',
+    'custom_event' => 'Custom Event',
+    'add' => 'Add',
+    'current_users' => 'Current Users',
+    'actions' => 'Actions',
+    'edit' => 'Edit',
+    'delete' => 'Delete',
+    'confirm_delete_user' => 'Are you sure you want to delete this user?',
+    'user_added_success' => 'User added successfully.',
+    'user_add_error' => 'Error occurred while adding user.',
+    'user_updated_success' => 'User updated successfully.',
+    'user_update_error' => 'Error occurred while updating.',
+    'user_deleted_success' => 'User deleted successfully.',
+    'user_delete_error' => 'Error occurred while deleting.',
+    'user_exists' => 'Username already exists.',
+    'all_fields_required' => 'Please enter all fields.',
+    'leave_empty' => 'Leave empty to not change',
+    'save_changes' => 'Save Changes',
+    'cancel' => 'Cancel',
+
+    // RSVP Page
+    'welcome_guest' => 'Welcome',
+    'dear_guest' => 'Dear Guest',
+    'guest_count' => 'Guest Count',
+    'table_number' => 'Table Number',
+    'confirm_attendance' => 'Confirm Attendance',
+    'decline_attendance' => 'Decline Attendance',
+    'add_to_calendar' => 'Add to Calendar',
+    'share_invitation' => 'Share Invitation',
+    'get_directions' => 'Get Directions',
+    'download_qr' => 'Download QR',
+    'entry_card' => 'Entry Card',
+    'qr_code' => 'QR Code',
+    'show_at_entrance' => 'Show this code at entrance',
+    'already_confirmed' => 'Your attendance has been confirmed!',
+    'already_declined' => 'Your decline has been recorded',
+    'success_confirmed' => 'Your attendance has been confirmed successfully!',
+    'success_declined' => 'Thank you, your decline has been recorded.',
+    'error_occurred' => 'An error occurred, please try again',
+    'invalid_link' => 'Invalid invitation link',
+    'csrf_error' => 'Security error, please reload the page',
+    'rate_limit_error' => 'Too many requests, please wait',
+    'connection_error' => 'Connection error',
+    'countdown_title' => 'Event Countdown',
+    'days' => 'Days',
+    'hours' => 'Hours',
+    'minutes' => 'Minutes',
+    'seconds' => 'Seconds',
+    'event_time_reached' => '🎉 Event Time! 🎉',
+    'enjoy_time' => 'Have a wonderful time!'
+];
