+++ public/css/login.css
@@ -1,0 +1,153 @@
+:root {
+    --primary-green: #9CAC80;
+    --dark-green: #7a8a5e;
+    --premium-black: #0a0a0a;
+}
+
+body {
+    font-family: 'Cairo', sans-serif;
+    background:
+        radial-gradient(circle at 20% 80%, rgba(156, 172, 128, 0.12) 0%, transparent 50%),
+        radial-gradient(circle at 80% 20%, rgba(160, 124, 94, 0.08) 0%, transparent 50%),
+        radial-gradient(circle at 40% 40%, rgba(212, 175, 55, 0.06) 0%, transparent 50%),
+        linear-gradient(135deg,
+            rgba(10, 10, 10, 0.98) 0%,
+            rgba(20, 20, 20, 0.95) 25%,
+            rgba(15, 15, 15, 0.97) 50%,
+            rgba(25, 25, 25, 0.94) 75%,
+            rgba(10, 10, 10, 0.98) 100%);
+    min-height: 100vh;
+    background-attachment: fixed;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem;
+}
+
+.login-container {
+    background: rgba(10, 10, 10, 0.5);
+    border: 1px solid rgba(255, 255, 255, 0.1);
+    border-radius: 20px;
+    padding: 2.5rem;
+    backdrop-filter: blur(10px);
+    width: 100%;
+    max-width: 420px;
+}
+
+.logo-icon {
+    width: 70px;
+    height: 70px;
+    background: linear-gradient(145deg, #ffffff 0%, #f8f9fa 100%);
+    border-radius: 18px;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 1rem;
+    box-shadow: 0 10px 30px rgba(255, 255, 255, 0.1);
+}
+
+label {
+    display: block;
+    margin-bottom: 0.5rem;
+    font-weight: 600;
+    color: #FFFFFF !important;
+    font-size: 0.875rem;
+}
+
+.form-input {
+    width: 100%;
+    padding: 0.75rem 1rem;
+    border-radius: 0.5rem;
+    background: rgba(255, 255, 255, 0.1);
+    color: white;
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    transition: all 0.3s ease;
+}
+
+.form-input::placeholder {
+    color: #9ca3af; /* text-gray-400 */
+}
+
+.form-input:focus {
+    outline: none;
+    border-color: var(--primary-green);
+    background-color: rgba(255, 255, 255, 0.15);
+    box-shadow: 0 0 0 3px rgba(156, 172, 128, 0.3);
+}
+
+.password-toggle {
+    position: absolute;
+    inset-inline-end: 0.75rem;
+    top: 50%;
+    transform: translateY(-50%);
+    color: #9ca3af;
+    cursor: pointer;
+}
+
+.btn-primary {
+    background: linear-gradient(135deg, var(--primary-green) 0%, var(--dark-green) 100%);
+    color: white;
+    border: none;
+    padding: 0.75rem 1.5rem;
+    border-radius: 50px;
+    font-weight: 600;
+    transition: all 0.3s ease;
+    width: 100%;
+}
+
+.btn-primary:hover:not(:disabled) {
+    transform: translateY(-2px);
+    box-shadow: 0 10px 25px rgba(156, 172, 128, 0.3);
+}
+
+.btn-primary:disabled {
+    opacity: 0.7;
+    cursor: not-allowed;
+}
+
+.message {
+    padding: 1rem;
+    border-radius: 0.75rem;
+    margin-bottom: 1.5rem;
+    font-size: 0.875rem;
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+}
+
+.message.success {
+    background-color: rgba(74, 222, 128, 0.1);
+    color: #a7f3d0;
+    border: 1px solid rgba(74, 222, 128, 0.3);
+}
+
+.message.error {
+    background-color: rgba(248, 113, 113, 0.1);
+    color: #fca5a5;
+    border: 1px solid rgba(248, 113, 113, 0.3);
+}
+
+.language-toggle button {
+    background: rgba(255, 255, 255, 0.1);
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    color: white;
+    padding: 0.5rem 1rem;
+    border-radius: 0.5rem;
+    font-size: 0.875rem;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    backdrop-filter: blur(10px);
+}
+
+.language-toggle button:hover {
+    background: rgba(255, 255, 255, 0.2);
+    border-color: var(--primary-green);
+}
+
+.attempts-counter {
+    text-align: center;
+    margin-bottom: 1rem;
+    font-size: 0.8rem;
+    color: #fca5a5; /* Light red */
+    font-weight: 600;
+}
