+++ views/layouts/admin_footer.php
@@ -1,0 +1,7 @@
+<footer class='text-center text-gray-600 text-sm mt-12 pt-8 border-t border-gray-200'>
+        <p><?= $lang === 'ar' ? 'جميع الحقوق محفوظة &copy; 2025 وصول' : 'All Rights Reserved &copy; 2025 Wosuol' ?></p>
+    </footer>
+
+    <script src="js/admin.js"></script>
+</body>
+</html>
